+++ src/library/renderhud/SurfaceARGB.cpp
@@ -22,6 +22,8 @@
 
 #include "../logging.h"
 
+#include <cmath>
+
 namespace libtas {
 
 SurfaceARGB::SurfaceARGB(int width, int height)
@@ -55,6 +57,108 @@ void SurfaceARGB::fillBorder(Color color, int t)
     }
 }
 
+void SurfaceARGB::drawLine(Color color, bool positive_slope)
+{ 
+    /* Special case of line along one axis */
+    if ((h == 1) || (w == 1)) {
+        fill(color);
+        return;
+    }
+    
+    /* Pre-compute color values */ 
+    uint32_t c = colorToValue(color);
+    uint32_t orig_color = c & 0xffffff;
+    uint32_t orig_alpha = color.a;
+ 
+    /* Write end-line pixels */
+    if (positive_slope) {
+        pixels[0] = c;
+        pixels[w*h-1] = c;
+    }
+    else {
+        pixels[w-1] = c;
+        pixels[w*(h-1)] = c;
+    }
+ 
+    const bool steep = h > w;
+    const double gradient = steep ? ((double)(w-1)/(double)(h-1)) : ((double)(h-1)/(double)(w-1));
+ 
+    double intery = gradient;
+ 
+    if (steep) {
+        for (int y = 1; y < (h-1); y++) {
+            int x = std::floor(intery);
+            uint32_t ratio = 256.0*(intery - (double)x);
+            
+            if (positive_slope) {
+                pixels[y*w+x] = orig_color | (((orig_alpha * (256-ratio)) >> 8) << 24);
+                pixels[y*w+x+1] = orig_color | (((orig_alpha * ratio) >> 8) << 24);
+            }
+            else {
+                pixels[(h-1-y)*w+x] = orig_color | (((orig_alpha * (256-ratio)) >> 8) << 24);
+                pixels[(h-1-y)*w+x+1] = orig_color | (((orig_alpha * ratio) >> 8) << 24);             
+            }
+            
+            intery += gradient;
+        }
+    } else {
+        for (int x = 1; x < (w-1); x++) {
+            int y = std::floor(intery);
+            uint32_t ratio = 256*(intery - y);
+            
+            if (positive_slope) {
+                pixels[y*w+x] = orig_color | (((orig_alpha * (256-ratio)) >> 8) << 24);
+                pixels[(y+1)*w+x] = orig_color | (((orig_alpha * ratio) >> 8) << 24);
+            }
+            else {
+                pixels[y*w+(w-1-x)] = orig_color | (((orig_alpha * (256-ratio)) >> 8) << 24);
+                pixels[(y+1)*w+(w-1-x)] = orig_color | (((orig_alpha * ratio) >> 8) << 24);
+            }
+
+            intery += gradient;
+        }
+    }
+}
+
+void SurfaceARGB::drawEllipse(Color color)
+{
+    /* Pre-compute color values */ 
+    uint32_t c = colorToValue(color);
+    uint32_t orig_color = c & 0xffffff;
+    uint32_t orig_alpha = color.a;
+
+    double radiusX = (w-1)/2;
+    double radiusY = (h-1)/2;
+    double radiusX2 = radiusX * radiusX;
+    double radiusY2 = radiusY * radiusY;
+
+    double quarter = std::round(radiusX2 / std::sqrt(radiusX2 + radiusY2));
+    for (double dx = 0; dx <= quarter; dx++) {
+        double dy = radiusY * std::sqrt(1 - dx * dx / radiusX2);
+        uint32_t alpha = std::round(256*(dy - std::floor(dy)));
+
+        setPixel4(radiusX, radiusY, dx, std::floor(dy), orig_color | (((orig_alpha * alpha) >> 8) << 24));
+        setPixel4(radiusX, radiusY, dx, std::floor(dy) - 1, orig_color | (((orig_alpha * (256-alpha)) >> 8) << 24));
+    }
+
+    quarter = std::round(radiusY2 / std::sqrt(radiusX2 + radiusY2));
+    for(double dy = 0; dy <= quarter; dy++) {
+        double dx = radiusX * std::sqrt(1 - dy * dy / radiusY2);
+        uint32_t alpha = std::round(256*(dx - std::floor(dx)));
+
+        setPixel4(radiusX, radiusY, std::floor(dx), dy, orig_color | (((orig_alpha * alpha) >> 8) << 24));
+        setPixel4(radiusX, radiusY, std::floor(dx) - 1, dy, orig_color | (((orig_alpha * (256-alpha)) >> 8) << 24));
+    }
+}
+
+void SurfaceARGB::setPixel4(int x, int y, int dx, int dy, uint32_t c)
+{
+    pixels[(y+dy)*w+x+dx] = c;
+    pixels[(y+dy)*w+x-dx] = c;
+    pixels[(y-dy)*w+x+dx] = c;
+    pixels[(y-dy)*w+x-dx] = c;
+}
+ 
 void SurfaceARGB::blit(const SurfaceARGB* src, int x, int y)
 {
     /* Code taken from SDL 2 software blitting BlitRGBtoRGBPixelAlpha() */